requirements.txt
markdown>=3.7
xhtml2pdf>=0.2.17
htmldocx>=0.0.6
beautifulsoup4>=4.12.3
python-docx>=1.1.2
cryptography>=41.0.5,<44
dify_plugin>=0.4.0,<0.7.0
zope.event>=5.0
zope.interface>=6.0
python-pptx>=1.0.2
